Defined some basic lexical types for tokenizing Clarion.
Added the TokenType and Token classes which provide an abstract
representation of an atomic unit of Clarion.
Added a Lexeme class that gives the abstract and concrete
representation of an atomic unit of Clarion.
Left the regular expressions that will be used to parse Clarion
tokens in the Token class blank for now. Will flesh these out in
the next commit.