Lexical spec will now be read from XML at runtime.
Since defining the abstract syntax tree in code would be needlessly
complicated I decided it would be done in XML or a DSL and created
at runtime. Because this is a significant break from how tokens
worked I decided to move tokens into their own lexical specification
XML file.
Added a LexicalSpec that serializes lexical specifications at
runtime.
The Token class can now be serialized to and from XML.
Moved definition of EOL, EOF, and ParseError tokens to the Token
class as they are not specific to the Clarion language.
Added a DefaultLexer that tokenizes a programming language given
a lexical specification and input stream. Reads input line-by-line.
Added a ClarionLexer that extends DefaultLexer by loading the
Clarion lexical specification during construction.