Reworked how defining and lexing tokens works.
Extended the Token type with several different Token types that
can be defined in XML with the tag name. I originally didn't take
this approach because marker types are normally a "code smell".
I realized polymorphism might be required to implement a dynamic
parser, which made it a little easier to justify this pattern. But,
this pattern does simplify the definition of tokens, and since most
of the work of defining the Clarion language (and others, maybe
some day) is done in the XML specifications, this adds a very
positive abstraction to the application in general.
My final thoughts on this: I don't like this pattern for general
programming (unless the marker types have additional _behavior_),
but this pattern makes sense for cases where behavior is heavily
defined in XML (or other) runtime resources.