Completed implementation of the DefaultParser.
Refined the implementation of DefaultParser, there's still more
that could be done to make it cleaner and more clear, but in its
current state it's understandable enough to present.
Added a ParseNode class that is used to build abstract syntax
tress.
Added a public constructor to DefaultParser and a Parse method so
the Clarion Parser can be implemented.