Working on the grammar and preparing the CLI for a presentation.
As the CLI is mostly to demonstrate the output of the parser, I'm
adding some (somewhat contrived) tooling to help show how the
parser is working for a presentation of the repository.
Also trying to work out some things with the grammar, namely
ambiguous cases. In Clarion, the assignment operator is the same
as the equality operator so an assignment and expression are
ambiguous.